Fix issue causing recreation of permissions
The comments incorrectly identified the issue. The recreation was caused by the terraform not being able to calculate the client id until running the apply since the value comes from a data source dependent on a resource that was being changed. As we know the client_id in this case will not change, adding it to ignore_changes means the value wont be changed anyway and terraform no longer attempts to calculate it