EAHW-2492: Terraform branch instance (#25)
* add branch stage to main and validate

* remove branch stage

* add terraform apply for branch to new drone file

* use latest image

* change type to docker

* add apply action

* use terraform image

* put command on single line

* change directory

* add tf init

* remove brackets

* remove more brackets

* change plan to apply

* auto approve apply

* add tf plan

* remove inline vars

* add realm

* Fix issue causing recreation of permissions

The comments incorrectly identified the issue. The recreation was caused by the terraform not being able to calculate the client id until running the apply since the value comes from a data source dependent on a resource that was being changed. As we know the client_id in this case will not change, adding it to ignore_changes means the value wont be changed anyway and terraform no longer attempts to calculate it

* remove unused code

---------

Co-authored-by: claire-guest <claire.guest@digital.homeoffice.gov.uk>
Co-authored-by: Stephen Kaye <stephen.kaye@digital.homeoffice.gov.uk>